// isoft-fe/src/app/models/incoming-text.model.ts
export interface IncomingTextModel {
    id?: number;
    documentTypeId?: number;
    documentTypeName?: string;
    toDate?: string;
    unitName?: string;
    textSymbol?: string;
    dateText?: string;
    content?: string;
    signer?: string;
    branchId?: number;
    departmentId?: number;
    departmentName?: string;
    receiverId?: number;
    receiverName?: string;
    fileUrl?: string;
    file?: any;
    fileName?: string;
}